Bug fixed

+++ Models/PG/schema.py
@@ -81,7 +81,7 @@ class CreateMerchantWithdrawlSchma:
     bank_id: int
     bank_currency_id: int
     account_currency: str
-    withdrawal_amount: int
+    withdrawal_amount: float
 
 
 @dataclass
